+++ momentum/js/playList.js
@@ -1,0 +1,14 @@
+'use strict'
+ const playList = [
+    {      
+      title: 'XXXtentacion',
+      src: './music/XXX.mp3',
+      duration: '01:59'
+    },  
+    {      
+      title: 'Skyrim',
+      src: './music/skyrim.mp3',
+      duration: '03:57'
+    }
+  ]
+  export default playList;
