+++ momentum/style.css/owfont-regular.css
@@ -6,12 +6,7 @@
  * -------------------------- */
 @font-face {
   font-family: 'owfont';
-  /* src: url('../momentum/fonts/owfont-regular.eot');
-  src: url('../momentum/fonts/owfont-regular.eot') format('embedded-opentype'),
-	   url('../momentum/fonts/owfont-regular.woff') format('woff'),
-	   url('../momentum/fonts/owfont-regular.ttf') format('truetype'),
-	   url('../momentum/fonts/owfont-regular.svg') format('svg');
- */
+ 
 
    src: url('../fonts/owfont-regular.eot?v=1.0.0');
   src: url('../fonts/owfont-regular.eot?#iefix&v=1.0.0') format('embedded-opentype'),
